Use size_t instead of unsigned long
This avoids the "conflicting types for built-in function" warning
when compiling without -ffreestanding.

+++ include/string.h
@@ -21,19 +21,21 @@
 #ifndef STRING_H
 #define STRING_H
 
+#include <stddef.h>
+
 /* string routines */
 
 #if !(USE_STATIC_INLINES)
 char *strcpy(char *dest, const char *src);
 #endif
 
-unsigned long strlcpy(char *dest,const char *src,unsigned long count);
-unsigned long int strlen(const char *s);
+size_t strlcpy(char *dest,const char *src,size_t count);
+size_t strlen(const char *s);
 short strlencpy(char *dest, const char *src);
 char *strcat(char *dest, const char *src);
 int strcmp(const char *a, const char *b);
-int strncmp(const char *a, const char *b, unsigned long int n);
-int strncasecmp(const char *a, const char *b, unsigned long int n);
+int strncmp(const char *a, const char *b, size_t n);
+int strncasecmp(const char *a, const char *b, size_t n);
 char *strchr(const char *s, int c);
 int toupper(int c);
 int sprintf(char *str, const char *fmt, ...);
@@ -53,24 +55,24 @@ static __inline__ char *strcpy(char *dest, const char *src)
 
 /* block memory routines */
 
-int memcmp(const void * s1, const void * s2, unsigned long int n);
+int memcmp(const void * s1, const void * s2, size_t n);
 
 /* moves length bytes from src to dst. returns dst as passed.
  * the behaviour is undefined if the two regions overlap.
  */
 void * memcpy(void * dst, const void * src,
-              long unsigned int length);
+              size_t length);
 
 /* moves length bytes from src to dst, performing correctly
  * if the two regions overlap. returns dst as passed.
  */
 void * memmove(void * dst, const void * src,
-               long unsigned int length);
+               size_t length);
 
 /* fills with byte c, returns the given address. */
-void * memset(void *address, int c, long unsigned int size);
+void * memset(void *address, int c, size_t size);
 
 /* clear memory */
-void bzero(void *address, long unsigned int size);
+void bzero(void *address, size_t size);
 
 #endif /* STRING_H */
